# src/docs/swagger.yaml
openapi: '3.0.2'
info:
    version: 1.0.0
    title: LOCAL API
servers:
    - url: /api
...
paths:
    /auth/register:
        post:
            summary: Register user
            operationId: registerUser
            tags:
                - Authentication
            parameters:
                - $ref: '#/components/parameters/headerLanguage'
                - $ref: '#/components/parameters/headerPlatform'
                - $ref: '#/components/parameters/headerVersion'
                - $ref: '#/components/parameters/headerTimeZone'
            requestBody:
                required: true
                content:
                    application/json:
                        examples:
                            EMAIL:
                                value:
                                    name: "John Doe"
                                    email: "johndoe@yopmail.com"
                                    password: "Qwerty@123"
                            PHONE:
                                value:
                                    name: "John Doe"
                                    countryCode: "+91"
                                    phone: "1234567890"
                                    password: "Qwerty@123"
                        schema:
                            type: object
                            required:
                                - name
                                - password
                            properties:
                                name:
                                    type: string
                                    description: Full name of the user
                                email:
                                    type: string
                                    format: email
                                    description: Email address of the user
                                countryCode:
                                    type: string
                                    description: Country code of the user
                                phone:
                                    type: string
                                    description: Mobile number of the user
                                password:
                                    type: string
                                    description: password for user account 
            responses:
                200:
                    description: User and auth token
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/ApiResponse'
                                properties:
                                    data:
                                        type: object
                                        properties:
                                            isMFAEnabled:
                                                type: boolean
                                            token:
                                                type: string
                                            user:
                                                $ref: '#/components/schemas/User'
                400:
                    $ref: '#/components/responses/BadRequestError'
    /auth/request-otp:
        post:
            summary: Request otp
            operationId: requestOtp
            tags:
                - Authentication
            parameters:
                - $ref: '#/components/parameters/headerLanguage'
                - $ref: '#/components/parameters/headerPlatform'
                - $ref: '#/components/parameters/headerVersion'
                - $ref: '#/components/parameters/headerTimeZone'
            requestBody:
                required: true
                content:
                    application/json:
                        examples:
                            EMAIL:
                                value:
                                    verifyType: 'EMAIL'
                                    otpType: 'FORGOT_PASSWORD'
                                    email: 'johndoe@yopmail.com'
                            PHONE:
                                value:
                                    verifyType: 'PHONE'
                                    otpType: 'FORGOT_PASSWORD'
                                    countryCode: '+91'
                                    phone: '1234567890'
                            BOTH:
                                value:
                                    verifyType: 'BOTH'
                                    otpType: 'FORGOT_PASSWORD'
                                    email: 'johndoe@yopmail.com'
                                    countryCode: '+91'
                                    phone: '1234567890'
                        schema:
                            type: object
                            required:
                                - verifyType
                                - otpType
                            properties:
                                verifyType:
                                    type: string
                                    description: Available values are EMAIL, PHONE, BOTH
                                otpType:
                                    type: string
                                    description: Available values are FORGOT_PASSWORD, CHANGE_PHONE, CHANGE_EMAIL, LOGIN
                                email:
                                    type: string
                                    description: Email address of User
                                countryCode:
                                    type: string
                                    description: Country code of user
                                phone:
                                    type: string
                                    description: Phone number of user
            responses:
                200:
                    description: Message indicating OTP has been sent successfully
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/ApiResponse'
                400:
                    $ref: '#/components/responses/BadRequestError'
    /auth/verify-otp:
        post:
            summary: Verify otp
            operationId: verifyOtp
            tags:
                - Authentication
            parameters:
                - $ref: '#/components/parameters/headerLanguage'
                - $ref: '#/components/parameters/headerPlatform'
                - $ref: '#/components/parameters/headerVersion'
                - $ref: '#/components/parameters/headerTimeZone'
            requestBody:
                required: true
                content:
                    application/json:
                        examples:
                            EMAIL:
                                value:
                                    verifyType: 'EMAIL'
                                    otpType: 'FORGOT_PASSWORD'
                                    email: 'johndoe@yopmail.com'
                                    emailToken: '123456'
                            PHONE:
                                value:
                                    verifyType: 'PHONE'
                                    otpType: 'FORGOT_PASSWORD'
                                    countryCode: '+91'
                                    phone: '1234567890'
                                    phoneToken: '123456'
                            BOTH:
                                value:
                                    verifyType: 'BOTH'
                                    otpType: 'FORGOT_PASSWORD'
                                    email: 'johndoe@yopmail.com'
                                    countryCode: '+91'
                                    phone: '1234567890'
                                    emailToken: '123456'
                                    phoneToken: '123456'
                        schema:
                            type: object
                            required:
                                - verifyType
                                - otpType
                            properties:
                                verifyType:
                                    type: string
                                    description: Available values are EMAIL, PHONE, BOTH
                                otpType:
                                    type: string
                                    description: Available values are FORGOT_PASSWORD, CHANGE_PHONE, CHANGE_EMAIL
                                email:
                                    type: string
                                    description: email address of user
                                countryCode:
                                    type: string
                                    description: Country code of user
                                phone:
                                    type: string
                                    description: Phone number of user
                                emailToken:
                                    type: string
                                    description: OTP received on email address of user
                                phoneToken:
                                    type: string
                                    description: OTP received on phone number of user
            responses:
                200:
                    description: Message indicating OTP has been verified successfully
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/ApiResponse'
                400:
                    $ref: '#/components/responses/BadRequestError'
...
    /auth/delete-account:
        delete:
            summary: Delete user account
            operationId: deleteMyAccount
            tags:
                - Authentication
            security:
                - ApiKeyAuth: []
            parameters:
                - $ref: '#/components/parameters/headerLanguage'
                - $ref: '#/components/parameters/headerPlatform'
                - $ref: '#/components/parameters/headerVersion'
                - $ref: '#/components/parameters/headerTimeZone'
            responses:
                200:
                    description: Account deleted successfully
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/ApiResponse'
                400:
                    $ref: '#/components/responses/BadRequestError'
...